fix: token 입력 및 제출하는 부분 클라이언트 컴포넌트로 분리
